refactor: move upload data protocol
